Moving export editor
