+++ src/main/transfer/Resources/modules/export/editor/components/main.jsx
@@ -5,10 +5,10 @@ import {PropTypes as T} from 'prop-types'
 import {trans} from '#/main/app/intl'
 import {Editor} from '#/main/app/editor/components/main'
 import {selectors} from '#/main/transfer/tools/export/store'
-import {ExportEditorFormat} from '#/main/transfer/tools/export/editor/components/format'
-import {ExportEditorActions} from '#/main/transfer/tools/export/editor/components/actions'
-import {ExportEditorPlaning} from '#/main/transfer/tools/export/editor/components/planing'
-import {ExportEditorOverview} from '#/main/transfer/tools/export/editor/components/overview'
+import {ExportEditorFormat} from '#/main/transfer/export/editor/components/format'
+import {ExportEditorActions} from '#/main/transfer/export/editor/components/actions'
+import {ExportEditorPlaning} from '#/main/transfer/export/editor/components/planing'
+import {ExportEditorOverview} from '#/main/transfer/export/editor/components/overview'
 
 const ExportEditor = (props) => {
   const entity = typeof props.formData.action !== 'undefined' ? props.formData.action.substring(0, props.formData.action.indexOf('_')) : ''
